Correction vector estimation

Batch effect estimation using an extended Kalman filter

EkfBE(
  refBatch,
  queBatch,
  pairs,
  sampling = FALSE,
  numSamples = NULL,
  verbose = FALSE
)

Arguments

refBatch

Reference batch.

queBatch

Query batch.

pairs

A numerical matrix containing MNNs pairs cell indexes. First column corresponds to query batch cells.

sampling

Sample MNNs pairs.

numSamples

If sampling, number of MNNs pairs samples to use on the estimation process.

verbose

Print output.

Value

A list containing the estimated correction vector and the estimation data. The length of the correction vector is equal to the number of genes.

Details

The input batches must have the same number of genes. The model used on the estimation has the form of g_ref = g_que + be, where the batch effect is represented as a value added to the reference gene expression, causing a linear deviation between the reference and the query batches.